==> notebooks/00_read_data.ipynb <==
Parámetros:


*   Data inputs
*   Columnas IDs



file = "C:/Users/Sebastian/Desktop/LOOKALIKE/lookalike_models/notebooks/data/endei_clu.csv"
id_cols = []

import numpy as np
import pandas as pd
import datetime
import os
import sys
import lkmodels

#### Import data

print(f'START --- {datetime.datetime.now()}')

df = pd.read_csv(file)
df.head()

df_seed = df[df.vd == True].drop('vd', axis=1)
df_poblacion = df[df.vd == False].drop('vd', axis=1)

#### Explore


print(df_poblacion.columns)

print(df_seed.shape)
print(df_poblacion.shape)

print(set(df_seed.columns) - set(df_poblacion.columns))
print(set(df_poblacion.columns) - set(df_seed.columns))

Validar que la semilla tenga menos observaciones que la población. Esto es una restricción de la función de bagging por el momento.

print(df_seed.shape[0] < df_poblacion.shape[0])

#### Transformaciones

common_columns = set(df_seed.columns).intersection(set(df_poblacion.columns))
# keep columnas en comun y numericas
X_seed = df_seed[common_columns].select_dtypes(['number'])
X_poblacion = df_poblacion[common_columns].select_dtypes(['number'])
# add ID_VARS si quedaron fuera
# if not set(X_seed.columns).intersection(set(ID_COLS)):
#   X_seed = pd.contact([df_seed[ID_COLS], X_seed], axis = 1)
# if not set(X_seed.columns).intersection(set(ID_COLS)):
#   X_seed = pd.contact([df_seed[ID_COLS], X_seed], axis = 1)

#### Save

# save
if not os.path.exists('data'):
  os.makedirs('data')
X_seed.to_pickle('data/X_seed.pkl')
X_poblacion.to_pickle('data/X_poblacion.pkl')

print(f'END --- {datetime.datetime.now()}')
