==> notebooks/01_transform_data.ipynb <==
ID_COLS = []
SEMILLA = 999

import numpy as np
import pandas as pd
import datetime

from sklearn.experimental import enable_iterative_imputer
from sklearn.impute import IterativeImputer

### Read

print(f'START --- {datetime.datetime.now()}')
      
X_seed = pd.read_pickle("data/X_seed.pkl")
X_poblacion = pd.read_pickle("data/X_poblacion.pkl")

### Explore

X_poblacion.head()

print(X_poblacion.isnull().mean())

print(X_seed.isnull().mean())

### Transforma

### Imputo Missings

### ver
### https://scikit-learn.org/stable/modules/generated/sklearn.impute.IterativeImputer.html

# cols_to_impute - [c for c in X_poblacion.columns if c not in ID_COLS]

# fit_transform y seed por separado
# imp = IterativeImputer(max_iter = 20, random_state = SEMILLA)
# X_poblacion[cols_to_impute] = imp.fit_transform(
#     X_poblacion[cols_to_impute])
# X_seed[cols_to_impute] = imp.fit_transform(
#     X_seed[cols_to_impute])

X_poblacion = X_poblacion.fillna(0)
X_seed = X_seed.fillna(0)

### Si hace falta tambien reemplazo infinitos 
X_poblacion.replace([np.inf, -np.inf], 0, inplace=True)
X_seed.replace([np.inf, -np.inf], 0, inplace=True)


### Save

# save (overwrite)
X_seed.to_pickle('data/X_seed.pkl')
X_poblacion.to_pickle('data/X_poblacion.pkl')

print(f'END --- {datetime.datetime.now()}')
