==> notebooks/02_score.ipynb <==
Este script corre el score lookalike de bagging en toda la poblacion.

Devuelve como resultado:
- un DataFrame con los scores
- un DataFrame con la importancia de variables promedio los clasificadores del bagging

ID_COLS = ['rut']
SEMILLA = 999
BAG_ITER = 100

import sys
# importar funcion bagging

import numpy as np
import pandas as pd
import datetime

from sklearn.tree import DecisionTreeClassifier
# from ... import bagged_clf 

### Data

print(f'START --- {datetime.datetime.now()}')

# read     
X_seed = pd.read_pickle("data/X_seed.pkl")
X_poblacion = read_pickle("data/X_poblacion.pkl")

### Run 

### Base Classifier
clf = DecisionTreeClassifier
clf_param = {'min_sample_leaf': 1, 'random_state': SEMILLA}

Corremos el algoritmo. Devuelve una lista de scores y un DataFrame de importancia de variables

### Lookalike scores
scores, feat_importance = bagged_clf(
    X_seed, X_poblacion, Classifier=clf,
    random_seed=SEMILLA, T=BAG_ITER, feat_importance=True,
    **clf_param)

### Save

# scores
df_scores = pd.DataFrame(ids_poblacion)
df_scores['score'] = scores
df_scores.to_csv('data/df_scores.csv', index=False)
# feature importance
feat_importance.to_csv('data/df_importance.csv', index=False)

print(f'START --- {datetime.datetime.now()}')
